don't safe position and thunkPosition in many cases where no change occured
In many cases position or thunkPosition--or both--have not
changed after a node's execution, so their value does not need
to be saved and restored.

The signature of the compile function has been changed so
that it now returns two pairs of flags, which tell the calling
function wether the sub node which has just been compiled might
change position or thunkPosition in "ok" and "ko" branches.

Depending on these values, and on the type of the current node,
it will be decided wether a {}-block with temporary variables
is wrapped around the definition of the sub node.

Because it is too late for the wrapping, after the sub
node has been compiled (the sub node's code has already
been written then), the flags are computed during a dry
compilation before the actual compilation (similar as in
https://github.com/pointlander/peg/commit/0c9235#L5R1202,
where it is done for a different purpose).

In one of the next revisions a switch will be added to turn
this feature off.